refactore: create modal stylex mixins. refactore modals stylec code
