--- src/components/Modals/User/UserRegisterModal/UserRegisterModal.js
@@ -1,32 +1,0 @@
-import React from "react";
-import "./userRegisterModal.scss";
-import Modal from "../../Common/Modal";
-import UsersModalDataWorker from "../UsersModalDataWorker";
-import { userModalTemplate } from "../ModalInputsTemplate";
-
-
-class UserRegisterModal extends React.Component {
-
-  render() {
-
-    const { closeModal, closeModalAndReload } = this.props;
-
-    return (
-      <Modal
-        dataWorker={
-          <UsersModalDataWorker
-            modalTemplate={userModalTemplate}
-            userData={null}
-            modalType='Register'
-            closeModal={closeModal}
-            closeModalAndReload={closeModalAndReload}
-          />
-        }
-        className="ModalBootstrap"
-        closeModal={closeModal}
-      />
-    );
-  }
-}
-
-export default UserRegisterModal;
