--- src/components/Modals/UserModal/UserModal.scss
@@ -1,18 +1,0 @@
-.ModalBootstrap {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  min-width: 52rem;
-}
-
-
-.modalBody{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 50rem;
-  height: auto;
-  background: rgba(255, 255, 255, 1);
-}
