+++ src/components/common/Styles/Mixins/Flex/flex.scss
@@ -1,0 +1,6 @@
+@mixin centerFlex {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
